queries for getInfo
I've uploaded the file as a text file since the rest of the code needs to be written, i.e. the GitHub token and using passing owner and name of the repositories as variables instead of hardcoded as it is in the queries. I've added comments to the queries to show what they are doing.